feat(lcd): Working touch calibration system with CST328
- Fixed CST328 driver to use correct 16-bit register addresses (0xD000)
- Fixed data format parsing (state field + proper coordinate extraction)
- Added TouchCalibration class for 9-point calibration
- Added serial command '<C>' to trigger calibration
- Touch coordinates now map correctly with piecewise linear interpolation
- Buttons respond correctly to touch input
- All debug output cleaned up

This commit preserves the working calibration system before simplification.

+++ lib/PicoDCCDisplay/touch_driver.h
@@ -61,10 +61,10 @@ private:
     // GPIO interrupt handler
     static void touchInterruptHandler(unsigned int gpio, uint32_t events);
     
-    // CST328 I2C communication
-    bool writeRegister(uint8_t reg, uint8_t value);
-    bool readRegister(uint8_t reg, uint8_t* value);
-    bool readMultipleRegisters(uint8_t reg, uint8_t* buffer, uint8_t length);
+    // CST328 I2C communication (CST328 uses 16-bit register addresses)
+    bool writeRegister(uint16_t reg, uint8_t value);
+    bool readRegister(uint16_t reg, uint8_t* value);
+    bool readMultipleRegisters(uint16_t reg, uint8_t* buffer, uint8_t length);
     
     // Hardware reset sequence
     void hardwareReset();
